+++ src/main/java/algo/mcts/MctsAlgo.java
@@ -32,8 +32,8 @@ public class MctsAlgo implements BaseAlgo {
     }
 
     @Override
-    public int nextMove(Board board) {
-        this.root = new MctsNode(null, -1, board, this.C);
+    public byte nextMove(Board board) {
+        this.root = new MctsNode(null, (byte) -1, board, this.C);
         for (int i = 0; i < this.epochs; i ++) {
             if (i % 10000 == 0) {
                 System.out.printf("Training epoch %d%n", i);
@@ -63,17 +63,17 @@ public class MctsAlgo implements BaseAlgo {
     }
 
     @Override
-    public List<Integer> getMovePredictions() {
+    public List<Byte> getMovePredictions() {
         assert this.root != null;
         return this.root.bestMoveSequence();
     }
 
     @Override
-    public int nextMoveWithTime(Board board, int time) {
+    public byte nextMoveWithTime(Board board, int time) {
         if (this.continueLastSearch) {
             this.setupRoot(board);
         } else {
-            this.root = new MctsNode(null, -1, board, this.C);
+            this.root = new MctsNode(null, (byte) -1, board, this.C);
         }
         System.gc();
         long startTime = System.currentTimeMillis();
@@ -86,7 +86,7 @@ public class MctsAlgo implements BaseAlgo {
 
     private void setupRoot(Board board) {
         if (this.root == null) {
-            this.root = new MctsNode(null, -1, board, this.C);
+            this.root = new MctsNode(null, (byte) -1, board, this.C);
         } else {
             this.continueSearch(board);
         }
@@ -96,7 +96,7 @@ public class MctsAlgo implements BaseAlgo {
         if (this.hasPonderedAfterSearch) {
             this.root = this.root.child(board);
             if (this.root == null) {
-                this.root = new MctsNode(null, -1, board, this.C);
+                this.root = new MctsNode(null, (byte) -1, board, this.C);
             }
             this.hasPonderedAfterSearch = false;
         } else {
@@ -131,7 +131,7 @@ public class MctsAlgo implements BaseAlgo {
 
     private void setupRootForPondering() {
         if (this.root == null) {
-            this.root = new MctsNode(null, -1, new Board(), this.C);
+            this.root = new MctsNode(null, (byte) -1, new Board(), this.C);
         } else {
             this.root = this.root.getBestChild();
         }
